+++ src/Randomizer/MapSpriteObject.js
@@ -177,7 +177,7 @@ class MapSpriteObject extends ZonePositionMixin(TableObject) {
         let equipment = [];
         equipmentOnce.forEach(item => {
             equipment.push(item);
-            if(!item.limitOne) {
+            if(!item.limitOne && !item.isWeapon) {
                 equipment.push(item);
                 equipment.push(item);
             }
@@ -214,6 +214,9 @@ class MapSpriteObject extends ZonePositionMixin(TableObject) {
             && !i.isKeyItem 
             && !i.isCondiment
             && !i.isBroken
+            && i.name !== "Ruler"       // Remove non-consumable items by default carried by Jeff
+            && i.name !== "Protractor"
+            && i.name !== "Show ticket" // Can be bought in the same room it is used
         );
         candidates = this.context.random.shuffleNormal(candidates, this.randomDegree || this.context.specs.randomDegree);
         chests = this.unassignedChests.sort((a, b) => a.caveRank - b.caveRank);